// FrontEnd/TeamHelperFront/src/app/api/api.ts
export interface Athlete {
    id?: string;
    name?: string | undefined;
    age?: number;
    phoneNumber?: string | undefined;
    email?: string | undefined;
    organization?: Organization;
    athleteIndicators?: AthleteIndicators[] | undefined;
}

export interface AthleteIndicators {
    id?: string;
    indicatorsData?: IndicatorsData[] | undefined;
    training?: Training;
    athlete?: Athlete;
    gearId?: string;
    gear?: Gear;
}
// ...
export interface Coach {
    id?: string;
    name?: string | undefined;
    age?: number;
    phoneNumber?: string | undefined;
    email?: string | undefined;
    organization?: Organization;
}
// ...
export interface Gear {
    id?: string;
    name?: string | undefined;
    athleteIndicators?: AthleteIndicators;
}

export interface GearListVM {
    gears?: GearLookupDTO[] | undefined;
}

export interface GearLookupDTO {
    id?: string;
    name?: string | undefined;
    athleteIndicators?: AthleteIndicators;
}

export interface GearVM {
    id?: string;
    name?: string | undefined;
    athleteIndicators?: AthleteIndicators;
}

export interface IndicatorsData {
    id?: string;
    pulse?: string | undefined;
    temperature?: string | undefined;
    bloodPressure?: string | undefined;
    athleteIndicators?: AthleteIndicators;
}
// ...
export interface Organization {
    id?: string;
    name?: string | undefined;
    adress?: string | undefined;
    phoneNumber?: string | undefined;
    email?: string | undefined;
    coaches?: Coach[] | undefined;
    athletes?: Athlete[] | undefined;
}
// ...
export interface Team {
    id?: string;
    name?: string | undefined;
    membNumber?: number;
    athlete?: Athlete[] | undefined;
    coach?: Coach;
}

export interface TeamListVM {
    teams?: TeamLookupDTO[] | undefined;
}

export interface TeamLookupDTO {
    id?: string;
    name?: string | undefined;
    membNumber?: number;
    athlete?: Athlete[] | undefined;
    coach?: Coach;
}

export interface TeamVM {
    id?: string;
    name?: string | undefined;
    membNumber?: number;
    athlete?: Athlete[] | undefined;
    coach?: Coach;
}

export interface Training {
    id?: string;
    date?: Date;
    location?: string | undefined;
    duration?: string | undefined;
    team?: Team;
    isInprocess?: boolean;
}

export interface TrainingListVM {
    trainings?: TrainingLookupDTO[] | undefined;
}

export interface TrainingLookupDTO {
    id?: string;
    date?: Date;
    location?: string | undefined;
    duration?: string | undefined;
    team?: Team;
    isInprocess?: boolean;
}

export interface TrainingVM {
    id?: string;
    date?: Date;
    location?: string | undefined;
    duration?: string | undefined;
    team?: Team;
    isInprocess?: boolean;
}